NAM-283 Split Express source code in separate modules
