+++ package.json
@@ -76,14 +76,14 @@
     "jszip": "^3.2.2",
     "locutus": "^2.0.11",
     "lodash-es": "^4.17.21",
-    "multer": "^1.4.2",
+    "multer": "^1.4.4",
     "ngx-color-picker": "^11.0.0",
     "ngx-filesaver": "^8.1.0",
     "pdf-lib": "^1.4.0",
     "pdfjs-dist": "^2.12.313",
-    "rgb-hex": "^3.0.0",
+    "rgb-hex": "^4.0.0",
     "rxjs": "~7.5.2",
-    "simple-excel-to-json": "^2.2.1",
+    "simple-excel-to-json": "^2.3.0",
     "trash": "^7.2.0",
     "tslib": "^2.3.1",
     "unzipper": "^0.10.10",
