+++ src-express/pdf/pdf-utils.ts
@@ -1,0 +1,4 @@
+
+export const adjustPointsForResults = (coordinate: number, change: number): number => {
+  return coordinate - (change);
+};
